move shared code to shared.rs
